@@ -1,63 +1,59 @@
 {
-    "name": "nino",
-    "description": "Moderation bot for Discord's hack week entry.",
-    "version": "0.5.2",
-    "main": "dist/bot.js",
-    "scripts": {
-        "build": "yarn lint && yarn clean && tsc",
-        "build:dev": "yarn clean && tsc -w",
-        "main": "yarn build && node dist/bot.js",
-        "dev": "nodemon dist/bot.js",
-        "lint": "eslint src --ext .ts --fix",
-        "clean": "rm -fr dist",
-        "test": "jest"
-    },
-    "maintainers": [
-        "August (auguwu)",
-        "dondish"
-    ],
-    "contributors": [
-        "Kyle (dvhe)",
-        "Wesselgame",
-        "ohlookitsderpy"
-    ],
-    "dependencies": {
-        "@augu/immutable": "0.0.8",
-        "@sentry/integrations": "5.8.0",
-        "@sentry/node": "5.9.0",
-        "common-tags": "1.8.0",
-        "eris": "0.11.1",
-        "express": "4.17.1",
-        "ioredis": "4.14.1",
-        "js-yaml": "3.13.1",
-        "mongoose": "5.7.13",
-        "ms": "2.1.2",
-        "pikmin": "0.1.13",
-        "prom-client": "11.5.3",
-        "wumpfetch": "0.3.1"
-    },
-    "devDependencies": {
-        "@types/common-tags": "1.8.0",
-        "@types/express": "4.17.2",
-        "@types/ioredis": "4.0.20",
-        "@types/jest": "24.0.23",
-        "@types/js-yaml": "3.12.1",
-        "@types/mongoose": "5.5.32",
-        "@types/ms": "0.7.31",
-        "@types/node": "12.12.14",
-        "@typescript-eslint/eslint-plugin": "2.9.0",
-        "@typescript-eslint/parser": "2.9.0",
-        "eslint": "6.7.1",
-        "jest": "24.9.0",
-        "nodemon": "2.0.1",
-        "ts-jest": "24.2.0",
-        "typescript": "3.7.2"
-    },
-    "jest": {
-        "transform": {
-            "^.+\\.tsx?$": "ts-jest"
-        },
-        "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
-        "moduleFileExtensions": ["ts", "tsx", "js", "jsx", "json"]
-    }
+  "name": "nino",
+  "description": "Moderation bot for Discord's hack week entry.",
+  "version": "0.7.0",
+  "main": "dist/index.js",
+  "scripts": {
+    "build:watch": "yarn lint && yarn test && yarn clean && tsc -w",
+    "build": "yarn lint && yarn test && yarn clean && tsc",
+    "clean": "rm -fr dist",
+    "main": "yarn build && node dist/index.js",
+    "lint": "eslint src --ext .ts --fix",
+    "test": "jest --config jest.config.js --forceExit",
+    "dev": "nodemon dist/index.js"
+  },
+  "maintainers": [
+    "August (Chris)",
+    "dondish"
+  ],
+  "contributors": [
+    "Kyle (dvhe)",
+    "Wesselgame",
+    "ohlookitsderpy"
+  ],
+  "dependencies": {
+    "@augu/immutable": "0.2.0",
+    "@sentry/node": "5.15.4",
+    "common-tags": "1.8.0",
+    "eris": "0.11.1",
+    "inversify": "5.0.1",
+    "inversify-inject-decorators": "3.1.0",
+    "ioredis": "4.16.1",
+    "js-yaml": "3.13.1",
+    "leeks.js": "0.0.8",
+    "mongoose": "5.9.7",
+    "ms": "2.1.2",
+    "prom-client": "11.5.3",
+    "reflect-metadata": "0.1.13",
+    "wumpfetch": "0.3.1"
+  },
+  "devDependencies": {
+    "@augu/eslint-config": "^1.3.3",
+    "@types/common-tags": "1.8.0",
+    "@types/express": "4.17.6",
+    "@types/ioredis": "4.14.9",
+    "@types/jest": "25.2.1",
+    "@types/js-yaml": "3.12.3",
+    "@types/mongoose": "5.7.10",
+    "@types/ms": "0.7.31",
+    "@types/node": "12.12.31",
+    "@typescript-eslint/eslint-plugin": "2.27.0",
+    "@typescript-eslint/parser": "2.27.0",
+    "eslint": "6.8.0",
+    "jest": "25.3.0",
+    "jest-mock-extended": "1.0.8",
+    "nodemon": "2.0.3",
+    "ts-jest": "25.3.1",
+    "typescript": "3.8.3"
+  }
 }
