@@ -1,57 +1,64 @@
-import { VERSION as __version__ } from 'eris';
+import { humanize, formatSize } from '../../util';
+import { VERSION as version } from 'eris';
+import { injectable, inject } from 'inversify';
 import { stripIndents } from 'common-tags';
-import { humanize } from '../../util';
-import NinoClient from '../../structures/Client';
+import { execSync } from 'child_process';
+import { TYPES } from '../../types';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
-import { execSync } from 'child_process';
+import Bot from '../../structures/Bot';
+import ts from 'typescript';
 
+const pkg = require('../../../package.json');
+
+@injectable()
 export default class StatisticsCommand extends Command {
-    constructor(client: NinoClient) {
-        super(client, {
-            name: 'statistics',
-            description: 'Gives you the bot\'s statistics',
-            aliases: [ 'stats', 'info', 'bot', 'botinfo' ],
-            category: 'Generic',
-            ownerOnly: false
-        });
-    }
+  constructor(
+    @inject(TYPES.Bot) client: Bot
+  ) {
+    super(client, {
+      name: 'statistics',
+      description: 'Gives you the bot\'s statistics',
+      aliases: ['stats', 'info', 'bot', 'botinfo'],
+      category: 'Generic',
+      ownerOnly: false,
+    });
+  }
 
-    getMostUsedCommand() {
-        if (Object.keys(this.client.stats.commandUsage).length > 0) {
-            const name = Object.keys(this.client.stats.commandUsage)
-            .map(key => ({ key, uses: this.client.stats.commandUsage[key].size })) // map key array to {key uses} array
-            .sort((a, b) => b.uses - a.uses) // Sort by uses
-            [0].key;
+  async run(ctx: Context) {
+    const { command, size: uses } = this.bot.statistics.getCommandUsages();
+    const build = await this.bot.database.getBuild();
+    const commit = execSync('git rev-parse HEAD').toString().trim();
+    const users = this.bot.client.guilds.reduce((a, b) => a + b.memberCount, 0);
+    const channels = Object.keys(this.bot.client.channelGuildMap).length;
+    const shardPing = this.bot.client.shards.reduce((a, b) => a + b.latency, 0);
+    const connection = await this.bot.database.admin.ping();
+    const memoryUsage = formatSize(process.memoryUsage().heapUsed);
 
-            return {
-                command: name,
-                size: this.client.stats.commandUsage[name].size,
-                users: this.client.stats.commandUsage[name].users.length
-            };  
-        }
-        return {command: 'None', size: 0, users: 0};
-    }
+    const embed = this.bot.getEmbed()
+      .setTitle(`${this.bot.client.user.username}#${this.bot.client.user.discriminator} | Realtime Statistics`)
+      .setDescription(stripIndents`
+        \`\`\`prolog
+        Guilds              => ${this.bot.client.guilds.size.toLocaleString()}
+        Users               => ${users.toLocaleString()}
+        Channels            => ${channels.toLocaleString()}
+        Shards              => ${ctx.guild!.shard.id}/${this.bot.client.shards.size} (${shardPing}ms avg.)
+        Uptime              => ${humanize(Date.now() - this.bot.client.startTime)}
+        Total Commands      => ${this.bot.manager.commands.size}
+        Messages Seen       => ${this.bot.statistics.messagesSeen.toLocaleString()}
+        Commands Executed   => ${this.bot.statistics.commandsExecuted.toLocaleString()}
+        Most Used Command   => ${command} (${uses} executions)
+        MongoDB Version     => v${build.version}
+        TypeScript Version  => v${ts.version}
+        Node.js Version     => ${process.version}
+        Eris Version        => ${version}
+        Nino Version        => ${pkg.version} (${commit.slice(0, 7)})
+        Database Connection => ${connection.ok === 1 ? 'Online' : 'Offline'}
+        Memory Usaage       => ${memoryUsage}
+        \`\`\`
+      `)
+      .build();
 
-    async run(ctx: Context) {
-        const command = this.getMostUsedCommand();
-        const build   = await this.client.database.getBuild();
-        const commit  = execSync('git rev-parse HEAD').toString().trim();
-        
-        return ctx.send(stripIndents`
-            \`\`\`prolog
-            Guilds              ~> ${this.client.guilds.size.toLocaleString()}
-            Users               ~> ${this.client.users.size.toLocaleString()}
-            Channels            ~> ${Object.keys(this.client.channelGuildMap).length.toLocaleString()}
-            Shards [C/T]        ~> [${ctx.guild.shard.id}/${this.client.shards.size}]
-            Uptime              ~> ${humanize(Date.now() - this.client.startTime)}
-            Commands            ~> ${this.client.manager.commands.size}
-            Messages Seen       ~> ${this.client.stats.messagesSeen.toLocaleString()}
-            Commands Executed   ~> ${this.client.stats.commandsExecuted.toLocaleString()}
-            Most Used Command   ~> ${command.command} (${command.size} executions)
-            Database Connection ~> v${build.version}
-            GitHub Commit       ~> ${commit.slice(0, 7)}
-            \`\`\`
-        `);
-    }
+    return ctx.embed(embed);
+  }
 }
