@@ -1,47 +1,51 @@
 import { Punishment, PunishmentType } from '../../structures/managers/PunishmentManager';
+import { injectable, inject } from 'inversify';
+import PermissionUtils from '../../util/PermissionUtils';
 import { Constants } from 'eris';
-import NinoClient from '../../structures/Client';
-import findUser from '../../util/UserUtil';
+import { TYPES } from '../../types';
+import findUser from '../../util/UserUtil'; 
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
-import PermissionUtils from '../../util/PermissionUtils';
+import Bot from '../../structures/Bot';
 
+@injectable()
 export default class KickCommand extends Command {
-    constructor(client: NinoClient) {
-        super(client, {
-            name: 'kick',
-            description: 'Kicks a user from the guild',
-            usage: '<user> <reason> [--reason]',
-            aliases: [ 'boot' ],
-            category: 'Moderation',
-            guildOnly: true,
-            botpermissions: Constants.Permissions.kickMembers,
-            userpermissions: Constants.Permissions.kickMembers
-        });
+  constructor(
+    @inject(TYPES.Bot) client: Bot
+  ) {
+    super(client, {
+      name: 'kick',
+      description: 'Kicks a user from the guild',
+      usage: '<user> <reason> [--reason]',
+      aliases: ['boot'],
+      category: 'Moderation',
+      guildOnly: true,
+      botPermissions: Constants.Permissions.kickMembers,
+      userPermissions: Constants.Permissions.kickMembers,
+    });
+  }
+
+  async run(ctx: Context) {
+    if (!ctx.args.has(0)) return ctx.send('No user has been specified');
+
+    const userID = ctx.args.get(0);
+    const user = findUser(this.bot, userID);
+    if (!user || user === undefined) return ctx.send('Cannot find member in this guild');
+
+    const member = ctx.guild!.members.get(user.id);
+    if (!member) return ctx.send(`User **\`${user.username}#${user.discriminator}\`** isn't in the guild`);
+    if (!PermissionUtils.above(ctx.message.member!, member)) return ctx.send('The user is above or the same heirarchy as you');
+
+    const reason = ctx.args.has(1) ? ctx.args.slice(1).join(' ') : undefined;
+    const punishment = new Punishment(PunishmentType.Kick, {
+      moderator: ctx.sender
+    });
+
+    try {
+      await this.bot.punishments.punish(member!, punishment, reason);
+      return ctx.send('User was successfully kicked');
+    } catch(e) {
+      ctx.send(`Unable to kick user: \`${e.message}\``);
     }
-
-    async run(ctx: Context) {
-        if (!ctx.args.has(0)) return ctx.send('You need to specify a user.');
-
-        const u = findUser(this.client, ctx.args.get(0))!;
-        if (!u) {
-            return ctx.send('I can\'t find this user!');
-        }
-        const member = ctx.guild.members.get(u.id);
-
-        if (!member) return ctx.send(`User \`${u.username}#${u.discriminator}\` is not in this guild?`);
-
-        if (!PermissionUtils.above(ctx.message.member!, member))
-            return ctx.send('The user is above you in the heirarchy.');
-
-            let reason = (ctx.flags.get('reason') || ctx.flags.get('r') || ctx.args.has(1) ? ctx.args.slice(1).join(' ') : false);
-            if (reason && typeof reason === 'boolean') return ctx.send('You will need to specify a reason');
-
-        const punishment = new Punishment(PunishmentType.Kick, {
-            moderator: ctx.sender
-        });
-
-        await ctx.send('User successfully kicked.');
-        await this.client.punishments.punish(member!, punishment, (reason as string | undefined));
-    }
+  }
 }
