@@ -1,51 +1,63 @@
 import { Punishment, PunishmentType } from '../../structures/managers/PunishmentManager';
+import { injectable, inject } from 'inversify';
+import PermissionUtils from '../../util/PermissionUtils';
 import { Constants } from 'eris';
-import NinoClient from '../../structures/Client';
-import findUser from '../../util/UserUtil';
+import { TYPES } from '../../types';
+import findUser from '../../util/UserUtil'; 
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
+import Bot from '../../structures/Bot';
 import ms = require('ms');
-import PermissionUtils from '../../util/PermissionUtils';
 
+@injectable()
 export default class MuteCommand extends Command {
-    constructor(client: NinoClient) {
-        super(client, {
-            name: 'mute',
-            description: 'Mutes a member from this guild',
-            usage: '<user> <reason> [--reason] [--time]',
-            aliases: ['slience'],
-            category: 'Moderation',
-            userpermissions: Constants.Permissions.manageRoles,
-            botpermissions: Constants.Permissions.manageRoles | Constants.Permissions.manageChannels
-        });
+  constructor(
+    @inject(TYPES.Bot) client: Bot
+  ) {
+    super(client, {
+      name: 'mute',
+      description: 'Mutes a member from this guild',
+      usage: '<user> <reason> | [time]',
+      aliases: ['slience'],
+      category: 'Moderation',
+      userPermissions: Constants.Permissions.manageRoles,
+      botPermissions: Constants.Permissions.manageRoles | Constants.Permissions.manageChannels,
+    });
+  }
+
+  async run(ctx: Context) {
+    if (!ctx.args.has(0)) return ctx.send('No user has been specified');
+
+    const userID = ctx.args.get(0);
+    const user = findUser(this.bot, userID);
+    if (!user || user === undefined) return ctx.send('Cannot find member in this guild');
+
+    const member = ctx.guild!.members.get(user.id);
+    if (!member) return ctx.send(`User **\`${user.username}#${user.discriminator}\`** isn't in the guild`);
+    if (!PermissionUtils.above(ctx.message.member!, member)) return ctx.send('The user is above or the same heirarchy as you');
+
+    const baseReason = ctx.args.has(1) ? ctx.args.slice(1).join(' ') : undefined;
+    let reason!: string;
+    let time!: string;
+  
+    if (baseReason) {
+      const sliced = baseReason.split(' | ');
+      reason = sliced[0];
+      time = sliced[1];
     }
 
-    async run(ctx: Context) {
-        if (!ctx.args.has(0)) return ctx.send('You need to specify a user.');
+    const t = time ? ms(time) : undefined;
+    const punishment = new Punishment(PunishmentType.Mute, {
+      moderator: ctx.sender,
+      temp: t
+    });
 
-        const u = findUser(this.client, ctx.args.get(0))!;
-        if (!u) return ctx.send('I can\'t find this user!');
-        const member = ctx.guild.members.get(u.id);
-
-        if (!member) return ctx.send(`User \`${u.username}#${u.discriminator}\` is not in this guild?`);
-
-        if (!PermissionUtils.above(ctx.message.member!, member)) return ctx.send('The user is above you in the heirarchy.');
-
-        let reason = (ctx.flags.get('reason') || ctx.flags.get('r') || ctx.args.has(1) ? ctx.args.slice(1).join(' ') : false);
-        if (reason && typeof reason === 'boolean') return ctx.send('You will need to specify a reason');
-
-        let time = (ctx.flags.get('time') || ctx.flags.get('t'));
-        if (typeof time === 'boolean') return ctx.send('You will need to specify time to be alloted');
-        
-        const t = !!time ? ms(time) : undefined;
-
-        const punishment = new Punishment(PunishmentType.Mute, {
-            moderator: ctx.sender,
-            temp: t
-        });
-        
-        await this.client.timeouts.cancelTimeout(member.id, ctx.guild, 'unmute');
-        await ctx.send('User successfully muted.');
-        await this.client.punishments.punish(member!, punishment, (reason as string | undefined));
+    await this.bot.timeouts.cancelTimeout(member.id, ctx.guild!, 'unmute');
+    try {
+      await this.bot.punishments.punish(member!, punishment, reason);
+      return ctx.send('User was successfully muted');
+    } catch (ex) {
+      return ctx.send(`Unable to mute user: \`${ex.message}\``);
     }
+  }
 }
