@@ -1,40 +1,43 @@
-import { Punishment, PunishmentType } from '../../structures/managers/PunishmentManager';
+import { injectable, inject } from 'inversify';
+import PermissionUtils from '../../util/PermissionUtils';
 import { Constants } from 'eris';
-import NinoClient from '../../structures/Client';
+import { TYPES } from '../../types';
 import findUser from '../../util/UserUtil';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
-import PermissionUtils from '../../util/PermissionUtils';
+import Bot from '../../structures/Bot';
 
+@injectable()
 export default class PardonCommand extends Command {
-    constructor(client: NinoClient) {
-        super(client, {
-            name: 'pardon',
-            description: 'Pardon a member from this guild',
-            usage: '<user> <amount>',
-            aliases: ['unwarn', 'forgive'],
-            category: 'Moderation',
-            userpermissions: Constants.Permissions.manageRoles | Constants.Permissions.manageChannels
-        });
-    }
+  constructor(
+    @inject(TYPES.Bot) client: Bot
+  ) {
+    super(client, {
+      name: 'pardon',
+      description: 'Pardon a member from this guild',
+      usage: '<user> <amount>',
+      aliases: ['unwarn', 'forgive'],
+      category: 'Moderation',
+      userPermissions: Constants.Permissions.manageRoles | Constants.Permissions.manageChannels
+    });
+  }
 
-    async run(ctx: Context) {
-        if (!ctx.args.has(0)) return ctx.send('You need to specify a user.');
-        if (!ctx.args.has(1) || !/^[0-9]+$/.test(ctx.args.get(1))) return ctx.send('Requires an amount of warnings to remove.');
+  async run(ctx: Context) {
+    if (!ctx.args.has(0)) return ctx.send('You\'ll need to specify a user');
+    if (!ctx.args.has(1) || !(/^[0-9]+$/).test(ctx.args.get(1))) return ctx.send('You didn\'t specify an amount of warnings to remove');
 
-        const u = findUser(this.client, ctx.args.get(0))!;
-        if (!u) return ctx.send('I can\'t find this user!');
-        const member = ctx.guild.members.get(u.id);
-        const amount = Number(ctx.args.get(1));
-        
+    const userID = ctx.args.get(0);
+    const user = findUser(this.bot, userID);
+    if (!user || user === undefined) return ctx.send('Cannot find member in this guild');
 
-        if (!member) return ctx.send(`User \`${u.username}#${u.discriminator}\` is not in this guild?`);
-
-        if (!PermissionUtils.above(ctx.message.member!, member)) return ctx.send('The user is above you in the heirarchy.');
-
-        await this.client.punishments.pardon(member!, amount);
-        const warns = await this.client.warnings.get(ctx.guild.id, member.id);
-        if (!warns) return ctx.send(`Successfully pardoned ${member.username}#${member.discriminator}! They now have 0 warnings!`);
-        return ctx.send(`Successfully pardoned ${member.username}#${member.discriminator}! They now have ${warns!.amount} warnings!`);
-    }
+    const member = ctx.guild!.members.get(user.id);
+    const amount = Number(ctx.args.get(1));
+    if (!member) return ctx.send(`User **\`${user.username}#${user.discriminator}\`** isn't in the guild`);
+    if (!PermissionUtils.above(ctx.message.member!, member)) return ctx.send('The user is above or the same heirarchy as you');
+  
+    await this.bot.punishments.pardon(member!, amount);
+    const warns = await this.bot.warnings.get(ctx.guild!.id, member.id);
+    if (!warns) return ctx.send(`${member.username}#${member.discriminator} now has no warnings`);
+    else return ctx.send(`Warned ${member.username}#${member.discriminator}! Now they have ${warns!.amount} warnings.`);
+  }
 }
