@@ -1,42 +1,49 @@
+import { injectable, inject } from 'inversify';
 import { Constants } from 'eris';
-import NinoClient from '../../structures/Client';
+import { TYPES } from '../../types';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
+import Bot from '../../structures/Bot';
 
+@injectable()
 export default class ReasonCommand extends Command {
-    constructor(client: NinoClient) {
-        super(client, {
-            name: 'reason',
-            description: 'Updates a case reason',
-            usage: '<caseID> <reason>',
-            aliases: [ 'update-reason' ],
-            category: 'Moderation',
-            guildOnly: true,
-            userpermissions: Constants.Permissions.banMembers,
-            botpermissions: Constants.Permissions.manageMessages
-        });
-    }
+  constructor(@inject(TYPES.Bot) client: Bot) {
+    super(client, {
+      name: 'reason',
+      description: 'Updates a case\'s reason',
+      usage: '<caseID> <reason>',
+      aliases: ['update'],
+      category: 'Moderation',
+      guildOnly: true,
+      userPermissions: Constants.Permissions.banMembers,
+      botPermissions: Constants.Permissions.manageMessages,
+    });
+  }
 
-    async run(ctx: Context) {
-        if (!ctx.args.has(0)) return ctx.send('Missing `<caseID>` argument');
-        if (!ctx.args.has(1)) return ctx.send('Missing `<reason>` argument');
+  async run(ctx: Context) {
+    if (!ctx.args.has(0)) return ctx.send('Missing the case ID to update');
+    if (!ctx.args.has(1)) return ctx.send('Missing the reason to update the case');
 
-        const caseID = ctx.args.get(0);
-        const reason = ctx.args.args.slice(1).join(' ');
-        const _case = await this.client.cases.get(ctx.guild.id, parseInt(caseID));
-        const settings = await this.client.settings.get(ctx.guild.id); 
+    const id = ctx.args.get(0);
+    const reason = ctx.args.args.slice(1).join(' ');
 
-        if (!_case || _case === null) return ctx.send(`Case #${caseID} was not found.`);
-        _case.reason = reason;
+    const _case = await this.bot.cases.get(ctx.guild!.id, parseInt(id));
+    const settings = await this.bot.settings.get(ctx.guild!.id);
 
-        await this.client.cases.update(ctx.guild.id, parseInt(caseID), {
-            $set: {
-                reason
-            }
-        }, async(error) => {
-            if (error) return ctx.send(`Unable to update case #${caseID}: \`${reason}\``);
-            const m = await this.client.getMessage(settings!.modlog, _case.message);
-            await this.client.punishments.editModlog(_case, m);
-        });
-    }
+    if (!_case || _case === null) return ctx.send(`Unable to find case **#${id}**, d-does it still exist?`);
+    _case.reason = reason;
+
+    await this.bot.cases.update(ctx.guild!.id, parseInt(id), {
+      $set: {
+        'reason': reason
+      }
+    }, async (error) => {
+      if (error) return ctx.send(`Unable to update case **#${id}**`);
+
+      const message = await this.bot.client.getMessage(settings!.modlog, _case.message);
+      await this.bot.punishments.editModlog(_case, message);
+
+      return ctx.send(`:ok_hand: Updated case **#${id}** for **${reason}**`);
+    });
+  }
 }
