@@ -1,36 +1,43 @@
-import { Punishment, PunishmentType } from '../../structures/managers/PunishmentManager';
+import { injectable, inject } from 'inversify';
 import { Constants } from 'eris';
-import NinoClient from '../../structures/Client';
 import findUser from '../../util/UserUtil';
+import { TYPES } from '../../types';
 import Command from '../../structures/Command';
 import Context from '../../structures/Context';
-import ms = require('ms');
-import PermissionUtils from '../../util/PermissionUtils';
+import Bot from '../../structures/Bot';
 
+@injectable()
 export default class WarnCommand extends Command {
-    constructor(client: NinoClient) {
-        super(client, {
-            name: 'warn',
-            description: 'Warns a member from this guild',
-            usage: '<user>',
-            aliases: ['addwarn'],
-            category: 'Moderation',
-            userpermissions: Constants.Permissions.manageRoles | Constants.Permissions.manageChannels
-        });
-    }
+  constructor(@inject(TYPES.Bot) client: Bot) {
+    super(client, {
+      name: 'warn',
+      description: 'Warns a member from this guild',
+      usage: '<user>',
+      aliases: ['addwarn'],
+      category: 'Moderation',
+      userPermissions:
+        Constants.Permissions.manageRoles |
+        Constants.Permissions.manageChannels,
+    });
+  }
 
-    async run(ctx: Context) {
-        if (!ctx.args.has(0)) return ctx.send('You need to specify a user.');
+  async run(ctx: Context) {
+    if (!ctx.args.has(0)) return ctx.send('You need to specify a user.');
 
-        const u = findUser(this.client, ctx.args.get(0))!;
-        if (!u) return ctx.send('I can\'t find this user!');
-        const member = ctx.guild.members.get(u.id);
+    const u = findUser(this.bot, ctx.args.get(0))!;
+    if (!u) return ctx.send('I can\'t find this user!');
+    const member = ctx.guild!.members.get(u.id);
 
-        if (!member) return ctx.send(`User \`${u.username}#${u.discriminator}\` is not in this guild?`);
+    if (!member) return ctx.send(`User \`${u.username}#${u.discriminator}\` is not in this guild?`);
 
-        const punishments = await this.client.punishments.addWarning(member!);
-        for (let i of punishments) await this.client.punishments.punish(member!, i, 'Automod');
-        const warns = await this.client.warnings.get(ctx.guild.id, member.id);
-        return ctx.send(`Successfully warned ${member.username}#${member.discriminator}! They now have ${warns!.amount} warnings!`);
-    }
+    const punishments = await this.bot.punishments.addWarning(member!);
+    for (let i of punishments)
+      try {
+        await this.bot.punishments.punish(member!, i, 'Automod');
+      } catch (e) {
+        return ctx.send(`Unable to punish, ${e.message}`);
+      }
+    const warns = await this.bot.warnings.get(ctx.guild!.id, member.id);
+    return ctx.send(`Successfully warned ${member.username}#${member.discriminator}! They now have ${warns === null ? 1 : warns.amount} warnings!`);
+  }
 }
