@@ -1,21 +1,34 @@
-import { Guild } from 'eris';
-import Client from '../structures/Client';
+import { injectable, inject } from 'inversify';
+import { Guild, TextChannel } from 'eris';
+import { TYPES } from '../types';
+import Client from '../structures/Bot';
 import Event from '../structures/Event';
 
+@injectable()
 export default class GuildLeftEvent extends Event {
-    constructor(client: Client) {
-        super(client, 'guildDelete');
-    }
+  constructor(
+    @inject(TYPES.Bot) client: Client
+  ) {
+    super(client, 'guildDelete');
+  }
 
-    async emit(guild: Guild) {
-        this.client.settings.remove(guild.id);
-        this.client.logger.log('discord', `Left ${guild.name} (${guild.id})`); // eslint-disable-line
-        this.client.editStatus('online', {
-            name: `${this.client.config['discord'].prefix}help | ${this.client.guilds.size.toLocaleString()} Guilds`,
-            type: 0
-        });
-        this.client.prom.guildCount.dec();
-        this.client.stats.guildCount--;
-        await this.client.redis.set('guilds', this.client.guilds.size);
+  async emit(guild: Guild) {
+    this.bot.settings.remove(guild.id);
+    this.bot.logger.warn(`Left guild ${guild.name} (${guild.id})`);
+
+    this.bot.status.updateStatus();
+    this.bot.prometheus.guildCount.inc();
+    this.bot.statistics.guildCount--;
+    await this.bot.redis.set('guilds', this.bot.client.guilds.size);
+    const channel = await this.bot.client.getRESTChannel('529593466729267200');
+    if (channel.type === 0) {
+      const chan = (channel as TextChannel);
+      const embed = this.bot.getEmbed()
+        .setAuthor(`| Joined ${guild.name} (${guild.id})`, undefined, this.bot.client.user.dynamicAvatarURL('png', 1024))
+        .setFooter(`Now at ${this.bot.client.guilds.size} Guilds`, this.bot.client.user.dynamicAvatarURL('png', 1024))
+        .build();
+
+      chan.createMessage({ embed });
     }
+  }
 }
