@@ -1,21 +1,30 @@
-import { Guild, Member } from 'eris';
-import Client from '../structures/Client';
-import Event from '../structures/Event';
 import { Punishment, PunishmentType } from '../structures/managers/PunishmentManager';
+import { injectable, inject } from 'inversify';
+import { Guild, Member } from 'eris';
+import { TYPES } from '../types';
+import Client from '../structures/Bot';
+import Event from '../structures/Event';
 
-export default class GuildMemberJoined extends Event {
-    constructor(client: Client) {
-        super(client, 'guildMemberAdd');
+@injectable()
+export default class GuildMemberJoinedEvent extends Event {
+  constructor(@inject(TYPES.Bot) client: Client) {
+    super(client, 'guildMemberAdd');
+  }
+
+  async emit(guild: Guild, member: Member) {
+    this.bot.automod.handleMemberJoin(member);
+
+    const cases = await this.bot.cases.getAll(guild.id);
+    const all = cases.sort(m => m.id).filter(m => m.victim === member.id);
+    if (all.length > 0) {
+      const latest = all[all.length - 1];
+      if (latest.type === 'mute') {
+        const punishment = new Punishment(PunishmentType.Mute, {
+          moderator: this.bot.client.user
+        });
+
+        await this.bot.punishments.punish(member, punishment, '[Automod] Mute Evading');
+      }
     }
-
-
-    async emit(guild: Guild, member: Member) {
-        this.client.autoModService.handleMemberJoin(member);
-        const cases = await this.client.cases.getAll(guild.id);
-        const his = cases.sort(m => m.id).filter(m => m.victim === member.id);
-        if (his.length > 0) {
-            const latest = his[his.length - 1];
-            if (latest.type === 'mute') await this.client.punishments.punish(member, new Punishment(PunishmentType.Mute, { moderator: guild.members.get(this.client.user.id)!.user}), 'mute evading');
-        }
-    }
+  }
 }
