@@ -1,12 +1,17 @@
-import Client from '../structures/Client';
+import { injectable, inject } from 'inversify';
+import { TYPES } from '../types';
+import Client from '../structures/Bot';
 import Event from '../structures/Event';
 
+@injectable()
 export default class ShardDisconnectedEvent extends Event {
-    constructor(client: Client) {
-        super(client, 'shardDisconnect');
-    }
+  constructor(
+    @inject(TYPES.Bot) client: Client
+  ) {
+    super(client, 'shardDisconnect');
+  }
 
-    async emit(id: number, error: any) {
-        this.client.logger.log('discord', `Shard #${id} has disconnected!\n${error}`);
-    }
+  async emit(error: any, id: number) {
+    this.bot.logger.warn(`Shard #${id} died`, error);
+  }
 }
