@@ -1,12 +1,17 @@
-import Client from '../structures/Client';
+import { injectable, inject } from 'inversify';
+import { TYPES } from '../types';
+import Client from '../structures/Bot';
 import Event from '../structures/Event';
 
+@injectable()
 export default class ShardReadyEvent extends Event {
-    constructor(client: Client) {
-        super(client, 'shardReady');
-    }
+  constructor(
+    @inject(TYPES.Bot) client: Client
+  ) {
+    super(client, 'shardReady');
+  }
 
-    async emit(id: number) {
-        this.client.logger.log('discord', `Shard #${id} is ready!`);
-    }
+  async emit(id: number) {
+    this.bot.logger.info(`Shard #${id} is connected to Discord!`);
+  }
 }
