@@ -1,15 +1,19 @@
-import NinoClient from '../Client';
-import AutoModSpam from '../automod/Spam';
-import AutoModInvite from '../automod/Invite';
+import { inject, injectable } from 'inversify';
 import { Message, Member } from 'eris';
-import AutoModBadWords from '../automod/Badwords';
-import AutoModRaid from '../automod/Raid';
-import AutoModDehoist from '../automod/Dehoisting';
-import AutoModMention from '../automod/Mention';
+import AutomodAccounts from '../automod/Accounts';
+import AutomodSwearing from '../automod/Badwords';
+import AutomodDehoist from '../automod/Dehoisting';
+import AutomodMention from '../automod/Mention';
+import AutomodInvite from '../automod/Invite';
+import AutomodSpam from '../automod/Spam';
+import AutomodRaid from '../automod/Raid';
+import { TYPES } from '../../types';
+import Bot from '../Bot';
+import 'reflect-metadata';
 
 /**
  * Service that generalizes automod event handling
- * 
+ *
  * @remarks
  * Automod features currently supported:
  * * anti-spam
@@ -18,42 +22,58 @@ import AutoModMention from '../automod/Mention';
  * * anti-raid
  * * auto dehoist
  * * auto mention
+ * * raid-member detection
  */
+@injectable()
 export default class AutomodService {
-    private spamhandler: AutoModSpam;
-    private invitehandler: AutoModInvite;
-    private badwordhandler: AutoModBadWords;
-    private raidhandler: AutoModRaid;
-    private dehoisthandler: AutoModDehoist;
-    private mentionhandler: AutoModMention;
+  private accounts: AutomodAccounts;
+  private swearing: AutomodSwearing;
+  private mentions: AutomodMention;
+  private invites: AutomodInvite;
+  private dehoist: AutomodDehoist;
+  private spam: AutomodSpam;
+  private raid: AutomodRaid;
 
-    constructor(client: NinoClient) {
-        this.spamhandler    = new AutoModSpam(client);
-        this.invitehandler  = new AutoModInvite(client);
-        this.badwordhandler = new AutoModBadWords(client);
-        this.raidhandler    = new AutoModRaid(client);
-        this.dehoisthandler = new AutoModDehoist(client);
-        this.mentionhandler = new AutoModMention(client);
-    }
+  constructor(
+    @inject(TYPES.Bot) bot: Bot
+  ) {
+    this.accounts = new AutomodAccounts(bot);
+    this.swearing = new AutomodSwearing(bot);
+    this.mentions = new AutomodMention(bot);
+    this.dehoist  = new AutomodDehoist(bot);
+    this.invites  = new AutomodInvite(bot);
+    this.spam     = new AutomodSpam(bot);
+    this.raid     = new AutomodRaid(bot);
+  }
 
-    /**
-     * Returns whether the event was handled
-     * 
-     * @param m the message
-     */
-    async handleMessage(m: Message): Promise<boolean> {
-        return await this.invitehandler.handle(m) || await this.badwordhandler.handle(m) || await this.spamhandler.handle(m) || await this.mentionhandler.handle(m);
-    }
+  /**
+   * Returns whether the event was handled
+   *
+   * @param m the message
+   */
+  async handleMessage(m: Message): Promise<boolean> {
+    return (
+      await this.invites.handle(m) ||
+      await this.swearing.handle(m) ||
+      await this.spam.handle(m) ||
+      await this.mentions.handle(m)
+    );
+  }
 
-    /**
-     * Returns whether the event was handled
-     * @param m the member
-     */
-    async handleMemberJoin(m: Member): Promise<boolean> {
-        return await this.raidhandler.handle(m) || await this.dehoisthandler.handle(m) || false;
-    }
+  /**
+   * Returns whether the event was handled
+   * @param m the member
+   */
+  async handleMemberJoin(m: Member): Promise<boolean> {
+    return (
+      await this.raid.handle(m) ||
+      await this.dehoist.handle(m) ||
+      await this.accounts.handle(m) || 
+      false
+    );
+  }
 
-    async handleMemberNameUpdate(m: Member): Promise<void> {
-        return this.dehoisthandler.handle(m);
-    }
+  async handleMemberNameUpdate(m: Member): Promise<void> {
+    return this.dehoist.handle(m);
+  }
 }
