@@ -1,58 +1,81 @@
-import { Message, Guild, TextChannel, User, EmbedOptions, Member } from 'eris';
-import MessageCollector from './MessageCollector';
+import { Message, TextChannel, User, EmbedOptions } from 'eris';
 import ArgumentParser from './parsers/ArgumentParser';
 import FlagParser from './parsers/FlagParser';
-import NinoClient from './Client';
+import Bot from './Bot';
 
 export interface DMOptions {
-    user: User;
-    content: string;
-    embed?: EmbedOptions;
+  user: User;
+  content: string;
+  embed?: EmbedOptions;
 }
 
 export default class CommandContext {
-    public client: NinoClient;
-    public message: Message;
-    public args: ArgumentParser;
-    public flags: FlagParser;
-    public collector: MessageCollector;
-    public guild: Guild;
-    public sender: User;
-    public me: Member;
+  public bot: Bot;
+  public message: Message;
+  public args: ArgumentParser;
+  public flags: FlagParser;
 
-    constructor(client: NinoClient, m: Message, args: string[]) {
-        Object.assign<this, Message>(this, m);
+  constructor(bot: Bot, m: Message, args: string[]) {
+    this.bot = bot;
+    this.message = m;
+    this.args = new ArgumentParser(args);
+    this.flags = new FlagParser(args);
+  }
 
-        this.client    = client;
-        this.message   = m;
-        this.args      = new ArgumentParser(args);
-        this.flags     = new FlagParser(args);
-        this.guild     = (m.channel as TextChannel).guild;
-        this.sender    = m.author;
-        this.collector = new MessageCollector(client);
-        this.me        = this.guild.members.get(client.user.id)!;
-    }
+  send(content: string) {
+    return this.message.channel.createMessage(content);
+  }
 
-    send(content: string) {
-        return this.message.channel.createMessage(content);
-    }
+  embed(content: EmbedOptions) {
+    return this.message.channel.createMessage({
+      embed: content,
+    });
+  }
 
-    embed(content: EmbedOptions) {
-        return this.message.channel.createMessage({
-            embed: content
-        });
-    }
+  code(lang: string, content: string) {
+    const cb = '```';
+    return this.send(`${cb}${lang}\n${content}${cb}`);
+  }
 
-    code(lang: string, content: string) {
-        const cb = '```';
-        return this.send(`${cb}${lang}\n${content}${cb}`);
-    }
+  get client() {
+    return this.bot.client;
+  }
 
-    async dm(options: DMOptions) {
-        const channel = await options.user.getDMChannel();
-        return channel.createMessage({
-            content: options.content,
-            embed: options.embed
-        });
-    }
+  get guild() {
+    return (this.message.channel instanceof TextChannel) ? 
+      (this.message.channel as TextChannel).guild : 
+      undefined;
+  }
+
+  get channel() {
+    return this.message.channel;
+  }
+
+  get sender() {
+    return this.message.author;
+  }
+
+  get member() {
+    return this.guild ? this.guild.members.get(this.sender.id) : undefined;
+  }
+
+  get me() {
+    return this.guild!.members.get(this.bot.client.user.id)!;
+  }
+
+  getSettings() {
+    return this.guild ? this.bot.settings.getOrCreate(this.guild.id) : null;
+  }
+
+  get redis() {
+    return this.bot.redis;
+  }
+
+  async dm(options: DMOptions) {
+    const channel = await options.user.getDMChannel();
+    return channel.createMessage({
+      content: options.content,
+      embed: options.embed,
+    });
+  }
 }
