@@ -1,30 +1,44 @@
 import { Message, TextChannel } from 'eris';
+import { replaceMessage } from '../../util';
 import PermissionUtil from '../../util/PermissionUtils';
-import NinoClient from '../Client';
+import Bot from '../Bot';
 
 export default class AutoModMention {
-    public client: NinoClient;
-    constructor(client: NinoClient) {
-        this.client = client;
+  public bot: Bot;
+  constructor(client: Bot) {
+    this.bot = client;
+  }
+
+  async handle(m: Message): Promise<boolean> {
+    const channel = m.channel as TextChannel;
+    const nino = channel.guild.members.get(this.bot.client.user.id)!;
+
+    if (
+      !PermissionUtil.above(nino, m.member!) ||
+      (m.channel as TextChannel)
+        .permissionsOf(m.author.id)
+        .has('manageMessages')
+    ) return false;
+
+    const settings = await this.bot.settings.get(channel.guild.id);
+    if (!settings || !settings.automod.mention) return false;
+
+    if (m.mentions.length >= 4) {
+      const punishments = await this.bot.punishments.addWarning(m.member!);
+      for (let punish of punishments) await this.bot.punishments.punish(
+        m.member!,
+        punish,
+        'Automod (Mention Spam)'
+      );
+
+      const response = !settings.responses.mention.enabled ?
+        replaceMessage('%author%, please don\'t mention more than 4 people at once!', m.author) :
+        replaceMessage(settings.responses.mention.message, m.author);
+
+      await m.channel.createMessage(response);
+      return true;
     }
 
-    async handle(m: Message): Promise<boolean> {
-        const channel = (m.channel as TextChannel);
-        const nino = channel.guild.members.get(this.client.user.id)!;
-
-        if (!PermissionUtil.above(nino, m.member!) || (m.channel as TextChannel).permissionsOf(m.author.id).has('manageMessages')) return false;
-
-        const settings = await this.client.settings.get(channel.guild.id);
-        if (!settings || !settings.automod.mention) return false;
-
-        
-        if (m.mentions.length >= 4) {
-            let punishment = await this.client.punishments.addWarning(m.member!);
-            for (let punish of punishment) await this.client.punishments.punish(m.member!, punish, 'Automod (Mention Spam)');
-            await m.channel.createMessage(`${m.member!.mention}, please don't mention more than 4 people at once!`);
-            return true;
-        }
-
-        return false;
-    }
+    return false;
+  }
 }
