@@ -1,53 +1,50 @@
 export default class ArgumentParser {
-    public args: string[];
+  public args: string[];
 
-    /**
-     * Constructs a new instance of the argument parser utility
-     * @param raw The raw arguments to use
-     */
-    constructor(raw: string[]) {
-        this.args = [];
-        
-        for (let i = 0; i < raw.length; i++) {
-            if (!raw[i].startsWith('--')) {
-                this.args.push(raw[i]);
-            } else {
-                break;
-            }
-        }
-    }
+  /**
+   * Constructs a new instance of the argument parser utility
+   * @param raw The raw arguments to use
+   */
+  constructor(raw: string[]) {
+    this.args = [];
 
-    /**
-     * Gets the argument by indexing
-     * @param i The index number to get from
-     */
-    get(i: number) {
-        return this.args[i];
+    for (let arg of raw) {
+      if (!arg.startsWith('--')) this.args.push(arg);
+      else break;
     }
+  }
 
-    /**
-     * Checks if the argument is empty
-     * @param i The index number
-     */
-    has(i: number) {
-        return !!this.args[i];
-    }
+  /**
+   * Gets the argument by indexing
+   * @param i The index number to get from
+   */
+  get(i: number) {
+    return this.args[i];
+  }
 
-    /**
-     * Gathers all arguments into a string
-     * @param sep The seperator
-     */
-    join(sep: string = ' ') {
-        return this.args.join(sep);
-    }
+  /**
+   * Checks if the argument is empty
+   * @param i The index number
+   */
+  has(i: number) {
+    return !!this.args[i];
+  }
 
-    /**
-     * Slices the argument from the `start` to `end`
-     * @param arg The argument number
-     * @param start The start of the slicing process
-     * @param end The end process
-     */
-    slice(start: number, end?: number) {
-        return this.args.slice(start, end);
-    }
+  /**
+   * Gathers all arguments into a string
+   * @param sep The seperator
+   */
+  join(sep: string = ' ') {
+    return this.args.join(sep);
+  }
+
+  /**
+   * Slices the argument from the `start` to `end`
+   * @param arg The argument number
+   * @param start The start of the slicing process
+   * @param end The end process
+   */
+  slice(start: number, end?: number) {
+    return this.args.slice(start, end);
+  }
 }
